+++ src/example/index.js
@@ -1,21 +1,15 @@
 import React from 'react'
-import { BrowserRouter, Match } from 'react-router'
-import Home from './pages/Home'
+import { Router } from 'react-router'
+import createBrowserHistory from 'history/createBrowserHistory'
 
+import { version } from '../../package.json'
+import Home from './pages/Home'
 import './index.css'
 
-const title = 'Reboo SlateJS Editor'
+const history = createBrowserHistory()
 
-const Example = () => {
-  return (
-    <BrowserRouter>
-      <div className="App">
-        {/* Begin Routes */}
-        <Match exactly pattern="/" render={() => <Home title={title} />} />
-        {/* End Routes */}
-      </div>
-    </BrowserRouter>
-  )
-}
-
-export default Example
+export default () => (
+  <Router history={history}>
+    <Home title='Nossas - SlateJS Editor' version={version} />
+  </Router>
+)