+++ src/package/plugins/slate-link-plugin/LinkButton.js
@@ -2,12 +2,13 @@ import React from 'react'
 import FontAwesome from 'react-fontawesome'
 import classnames from 'classnames'
 
+import { Button} from '../../components/button'
 import { insertLinkStrategy, hasLinks } from './LinkUtils'
 
-
-const LinkButton = ({ state, onChange, className, style }) => (
-  <button
+const LinkButton = ({ state, onChange, className, style, type }) => (
+  <Button
     style={style}
+    type={type}
     onMouseDown={e => onChange(insertLinkStrategy(state))}
     className={classnames(
       'slate-link-plugin--button',
@@ -16,7 +17,7 @@ const LinkButton = ({ state, onChange, className, style }) => (
     )}
   >
     <FontAwesome name="link" />
-  </button>
+  </Button>
 )
 
 export default LinkButton