+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="2.6.8"></a>
+## [2.6.8](https://github.com/nossas/slate-editor/compare/v2.6.7...v2.6.8) (2017-08-18)
+
+
+### Bug Fixes
+
+* **button lib:** Revert built lib files from Button component feature ([c3eb36f](https://github.com/nossas/slate-editor/commit/c3eb36f))
+* **example:** react-router router implementation fix [#37](https://github.com/nossas/slate-editor/issues/37) ([3e9eafa](https://github.com/nossas/slate-editor/commit/3e9eafa))
+
+
+
 <a name="2.6.7"></a>
 ## [2.6.7](https://github.com/nossas/slate-editor/compare/v2.6.6...v2.6.7) (2017-06-29)